refactor: refactoring of petition card info and hook
