+++ src/components/organisms/PetitionCard/PetitionCard.tsx
@@ -1,20 +1,21 @@
-import { useNavigate } from "react-router-dom";
 import { Icon } from "../../atoms/Icon/Icon";
 import S from "./PetitionCard.module.scss";
 import { PetitionStats } from "../../molecules/PetitionStats/PetitionStats";
+import { Petition } from "../../../utils/types";
+import { useNav } from "../../../hooks/useNav";
 
 type PetitionCardProps = {
-  id: number;
-  text: string;
+  data: Petition;
 };
-export const PetitionCard = ({ id, text }: PetitionCardProps) => {
-  const navigate = useNavigate();
-  const goToPetition = () => {
-    navigate(`/tickets/${id}`);
-  };
+
+export const PetitionCard = ({ data }: PetitionCardProps) => {
+  const { id, text, finish, theme, date } = data;
+
+  const { goToPetition } = useNav(id);
+
   return (
     <div onClick={goToPetition} className={S.card}>
-      <PetitionStats id={id} />
+      <PetitionStats theme={theme} date={date} finish={finish} id={id} />
       <div className={S.text}>
         <p>{text}</p>
       </div>
