+++ src/hooks/useNav.tsx
@@ -1,9 +1,15 @@
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 
-export const useNav = () => {
+export const useNav = (totalId?: number) => {
   const { id } = useParams();
+  const navigate = useNavigate();
+
+  const goToPetition = () => {
+    navigate(`/tickets/${totalId}`);
+  };
 
   return {
     id: Number(id),
+    goToPetition,
   };
 };
