+++ src/utils/types.ts
@@ -12,3 +12,19 @@ export type TemplateProps = {
 export type PetitionDefineType = {
   id: number;
 };
+
+export type ChatMessage = {
+  sender: string;
+  message: string;
+  timestamp: string;
+};
+
+export type Petition = {
+  theme: string;
+  id: number;
+  date: string;
+  text: string;
+  images: string[];
+  finish: boolean;
+  chat: ChatMessage[];
+};
